+++ nakadion/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [X]
+## [0.22.1] - 2020-05-05
+
+### CHANGED
+
+- Fixed event type partition tracking
 
 ### ADDED
 
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### REMOVED
 
-- `WarnStreamStalledSecs` removed
+- `WarnStreamStalledSecs` removed (this just affects logging)
 
 ## [0.22.0] - 2020-04-28